Extract pass-2 verification reconciliation into its own module
Moves 8 of the brief's 12 candidate functions to the new
src/main_app/verification_reconciliation.py, verbatim, with a module-level
import back into processing.py:
_apply_pass2_heuristic_rolls, _proposed_span_agrees, _corroborates_hold,
_corroborated_span, _exclude_kept_spans_from_verification,
_gate_verification_ads_by_confidence, _covered_by_cuts,
_drop_uncovered_pass2_ads.

Four functions from the brief's original list stay in processing.py, each
for a monkeypatch/introspection collision found empirically (full-suite
run), not assumed up front:

- _validate_verification_ads, _file_corroborated_hold_approvals: both call
  _build_validator / _load_user_corrections, which are also called by
  pass-1 code (_refine_and_validate, _build_recut_ad_list) that stays in
  processing.py. Moving the two shared helpers to avoid a reverse import
  breaks tests/unit/test_validate_verification_ads.py, which patches
  processing._build_validator / processing._load_user_corrections and
  expects the patch to reach these functions' internals; Python resolves
  unqualified names via the defining module's globals, not the importer's,
  so the patch would stop taking effect once the two functions moved.

- _pass2_cuts_in_original: calls get_replacement_duration(). Moving it
  broke tests/unit/test_pass2_asset_cuts.py, which patches
  processing.get_replacement_duration and expects that patch to reach this
  function; same root cause as above, discovered only after moving it and
  rerunning the suite.

- _recut_processed_audio: tests/unit/test_cut_partition.py has a source-scan
  test (inspect.getsource(processing) + regex) asserting every
  AudioProcessor.process_episode call site lives in processing.py; moving
  this function's call site out of the file drops the count below the
  test's threshold.

_build_validator and _load_user_corrections are untouched, staying exactly
where they were.

Zero test edits; full suite green (5345 passed). processing.py: 4428 -> 4109
lines (-319). New module: 344 lines.
